[Core] Treat temp file rename as a change event in file watcher
The TextFileUtility creates a temp file in the same directory as
the file being saved but changes the filename so it starts with '.#'
The native file watcher on Mac does not always generate a change
event when a file is saved using the TextFileUtility so we treat
a rename event of the temp file to the file being saved as a change
event that will be raised by FileService.FileChanged. This fixes
the SaveFileInProjectExternally test that was failing.