[Git] Stop using FileStatusTracker when working with stashes
The file watcher will now generate the required Changed and Deleted
events. Freezing and Thawing the FileService to get the same behaviour
as before when the FileService notifications were applied after the
stash was applied.

Behaviour is a bit broken here compared with before. The native file
watcher seems to want to generate - FileChanged, FileCreated and
FileDeleted. This results in any open documents that were changed
by the stash to be closed. Previously these would be reloaded. The
native file watcher gets a changed, created and deleted event flag
in the same message and then fires three events - changed, created
and deleted event even though the file exists.