[Core] Fix project not being reloaded when changed externally
Editing the .csproj file externally was not resulting in the project
being reloaded in the IDE. The FileService.FileChanged event was
used by the RootWorkspace before the FileStatusTracker processed
the same event so the NeedsReload was not true. As a workaround
the FileService.FileChanged event is registered by the workspace item
first.