[Core] RootWorkspace now uses ReloadRequired events
The RootWorkspace no longer uses the FileService.FileChanged to
detect a reload is need for solutions or projects. Instead it uses
the solution's ReloadRequired and ItemReloadRequired events. The
ReloadRequired event is fired by the FileStatusTracker when the
solution file is changed externally. The ItemReloadRequired event
is fired by the FileStatusTracker for any solution item if its file
is changed externally.